+++ exeris-kernel-community/src/test/java/eu/exeris/kernel/community/http/CommunityHttpSecurityAdmissionIntegrationTest.java
@@ -434,7 +434,8 @@ class CommunityHttpSecurityAdmissionIntegrationTest {
                 HttpConfig.DEFAULT_MAX_HEADER_SIZE,
                 HttpConfig.DEFAULT_MAX_REQUEST_BODY_BYTES,
                 false,
-                HttpVersion.HTTP_1_1
+                HttpVersion.HTTP_1_1,
+                "127.0.0.1" + ":" + port
         );
     }